New code

+++ src/OrderService/DTOs/OrderDto.cs
@@ -3,7 +3,7 @@ namespace OrderService.DTOs;
 public class OrderDto
 {
     public Guid Id { get; set; }
-    public decimal TotalPrice { get; set; }
+    public int TotalPrice { get; set; }
     public string Buyer { get; set; } = string.Empty;
     public string Seller { get; set; } = string.Empty;
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
@@ -12,7 +12,7 @@ public class OrderDto
 
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+    public int Price { get; set; }
     public string Category { get; set; } = string.Empty;
     public string ImageUrl { get; set; } = string.Empty;
     public int StockQuantity { get; set; }
