+++ src/SearchService/Models/Product.cs
@@ -2,7 +2,7 @@ using MongoDB.Entities;
 namespace SearchService;
 public class Product : Entity
 {
-    public decimal TotalPrice { get; set; }
+    public int TotalPrice { get; set; }
     public string Buyer { get; set; }
     public string Seller { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
@@ -11,7 +11,7 @@ public class Product : Entity
 
     public string Name { get; set; }
     public string Description { get; set; }
-    public decimal Price { get; set; }
+    public int Price { get; set; }
     public string Category { get; set; }
     public string ImageUrl { get; set; }
     public int StockQuantity { get; set; }
